# documentation/api/openapi/carehub-openapi.yaml
openapi: 3.1.0
info:
  title: Carehub API
  description: "API per gestione pazienti, medici, appuntamenti"
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-jwt: []
paths:
...
  /api/availabilities:
    get:
      tags:
      - availability-controller
      operationId: getAll
      parameters:
      - name: isActive
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableDate"
    put:
      tags:
      - availability-controller
      operationId: update_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailableSlotsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
    post:
      tags:
      - availability-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailableSlotsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
  /api/appointments/{id}:
    get:
      tags:
      - appointment-controller
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Appointment"
    put:
      tags:
      - appointment-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
    delete:
      tags:
      - appointment-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
...
  /api/appointments:
    get:
      tags:
      - appointment-controller
      operationId: getAll_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      tags:
      - appointment-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
  /api/appointments/filter:
    post:
      tags:
      - appointment-controller
      operationId: filter_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentFilter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
...
  /api/reports/appointment/{id}:
    get:
      tags:
      - medical-report-controller
      operationId: getByAppointmentId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MedicalReport"
...
  /api/availabilities/{id}:
    get:
      tags:
      - availability-controller
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AvailableDate"
...
  /api/availabilities/doctor/{doctorId}:
    get:
      tags:
      - availability-controller
      operationId: getByDoctor
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: isActive
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableDate"
  /api/availabilities/doctor/{doctorId}/date/{date}:
    get:
      tags:
      - availability-controller
      operationId: getSlotsByDoctorAndDate
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
      - name: isActive
        in: query
        required: false
        schema:
          type: boolean
      - name: notBooked
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAvailableSlotsRequest"
    delete:
      tags:
      - availability-controller
      operationId: deleteByDoctorAndDate
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
...
components:
  schemas:
...
    SimpleResult:
      type: object
      properties:
        result:
          type: string
        resultMessageHeader:
          type: string
        resultMessage:
          type: string
    CreateAvailableSlotsRequest:
      type: object
      properties:
        doctorId:
          type: integer
          format: int64
        availableDay:
          type: string
          format: date
        durationMinutes:
          type: integer
          format: int32
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotsDto"
    SlotsDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        time:
          type: string
        selected:
          type: boolean
        booked:
          type: boolean
    Appointment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientId:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
        availabilityId:
          type: integer
          format: int64
        visitType:
          type: string
          enum:
          - VISITA_CARDIOLOGICA
          - VISITA_ORTOPEDICA
          - VISITA_DERMATOLOGICA
        appointmentDay:
          type: string
          format: date
        appointmentTime:
          type: string
          example: 14:30:00
        status:
          type: string
          enum:
          - PRENOTATO
          - EFFETTUATA
        notes:
          type: string
        active:
          type: boolean
...
    MedicalReport:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientId:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
        appointmentId:
          type: integer
          format: int64
        summary:
          type: string
        notes:
          type: string
        cost:
          type: number
        fileName:
          type: string
        contentType:
          type: string
        reportDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AppointmentFilter:
      type: object
      properties:
        patientId:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
        visitType:
          type: string
          enum:
          - VISITA_CARDIOLOGICA
          - VISITA_ORTOPEDICA
          - VISITA_DERMATOLOGICA
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - PRENOTATO
          - EFFETTUATA
...
    AvailableDate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
        availableDay:
          type: string
          format: date
        availableTime:
          type: string
          example: 14:30:00
        createdAt:
          type: string
          format: date-time
        durationMinutes:
          type: string
          enum:
          - MIN_60
        active:
          type: boolean
        booked:
          type: boolean
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT